renamed and moved around some files, consolidated DB functions into 1 controller, added test data to NOAA API
