+++ server/controllers/apiController.js
@@ -2,43 +2,28 @@ const path = require('path');
 const fetch = require('node-fetch')
 //const mongoose = require('mongoose')
 require('dotenv').config();
-const {weatherData, Fips} = require('./apiModel');
-const {NOAA_KEY} = require('./secrets.js')
+const { weatherData } = require('../models/dbModel');
+const { NOAA_KEY } = require('../secrets.js');
 
 const apiController = {};
-
-//Retreiving FIPS code from MongoDB from client entry of county and state
-apiController.getFips = async(req, res, next) => {
-  try{
-    let {county, state} = req.params
-    const doc = await Fips.findOne({county: county, state: state})
-    res.locals.data = doc.fips;
-    return next();
-  } catch (error) {
-    return next({
-      log: 'apiController.getFips failed',
-      message: 'failed to retrieve FIPS code from database'
-    })
-    } 
-}
 
 // GET request for weather data at FIPS code passed in from prev middleware
 apiController.noaaData = async (req, res, next) => { 
   const fipsCode = res.locals.data;
-    await fetch(`https://www.ncei.noaa.gov/cdo-web/api/v2/data?enddate=2022-08-03&startdate=2012-08-03&locationid=FIPS:${fipsCode}&datasetid=GSOY&datatypeid=PSUN&datatypeid=TAVG&datatypeid=TMAX&datatypeid=TMIN&datatypeid=DX32&datatypeid=DX70&datatypeid=DX90&datatypeid=AWND&datatypeid=WSF2&limit=1000`,{
+  await fetch(`https://www.ncei.noaa.gov/cdo-web/api/v2/data?enddate=2022-08-03&startdate=2012-08-03&locationid=FIPS:${fipsCode}&datasetid=GSOY&datatypeid=PSUN&datatypeid=TAVG&datatypeid=TMAX&datatypeid=TMIN&datatypeid=DX32&datatypeid=DX70&datatypeid=DX90&datatypeid=AWND&datatypeid=WSF2&limit=1000`,{
     headers:{
-      'Token' : NOAA_KEY
+      'token' : NOAA_KEY
     }})
     .then(res => {console.log(res); return res.json(); })
-  .then(noaaData => {
-    res.locals.weather = noaaData;
-    return next();
-  })
-  .catch(err => next({
-    log: 'apiController.saveData failed',
-    message: 'failed to save weatherData to database'
-  }));
-}
+    .then(noaaData => {
+      res.locals.weather = noaaData;
+      return next();
+    })
+    .catch(err => next({
+      log: 'apiController.saveData failed',
+      message: 'failed to save weatherData to database'
+    }));
+};
