--- server/routes/contact_us.js
@@ -1,9 +1,0 @@
-"use strict"
-const express = require('express');
-const contactRouter = express.Router();
-
-
-
-
-
-module.exports = contactRouter;
